# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/user/login": {
      "post": {
        "description": "通过用户名/邮箱/手机号登录系统",
        "operationId": "UserController_login",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/loginUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "用户登录",
        "tags": [
          "用户模块"
        ]
      }
    },
    "/api/user/createUser": {
      "post": {
        "description": "",
        "operationId": "UserController_create",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "创建用户",
        "tags": [
          "用户模块"
        ]
      }
    },
    "/api/user": {
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户模块"
        ]
      }
    },
    "/api/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户模块"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户模块"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "用户模块"
        ]
      }
    },
    "/api/weixin": {
      "post": {
        "operationId": "WeixinController_create",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeixinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Weixin"
        ]
      },
      "get": {
        "operationId": "WeixinController_findAll",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Weixin"
        ]
      }
    },
    "/api/weixin/{id}": {
      "get": {
        "operationId": "WeixinController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Weixin"
        ]
      },
      "patch": {
        "operationId": "WeixinController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWeixinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Weixin"
        ]
      },
      "delete": {
        "operationId": "WeixinController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-custom-lang",
            "schema": {
              "example": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Weixin"
        ]
      }
    }
  },
  "info": {
    "title": "API项目",
    "description": "接口文档",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Nestjs Swagger",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "loginUserDto": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "description": "用户名",
            "example": "XXXXX"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "description": "密码",
            "example": "123456"
          },
          "phoneNumber": {
            "type": "string",
            "description": "手机号",
            "example": "13800138000"
          },
          "code": {
            "type": "string",
            "description": "验证码",
            "example": "123456"
          },
          "type": {
            "type": "string",
            "description": "类型",
            "example": "login"
          }
        },
        "required": [
          "type"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "CreateWeixinDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateWeixinDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}